# tpl/partials: Make sure a cached partial is invoked only once