@bep 2022-02-16
OK, I didn't get the new numbers from the "template hints" output to add up, so I adjusted the logic a little and added a test case.

@davidsneighbour @moorereason @jmooring some feedback on this would be appreciated, ref. my test output above:

```bash
 cumulative       average       maximum      cache  percent  cached  total
       duration      duration      duration  potential   cached   count  count  template
     ----------      --------      --------  ---------  -------  ------  -----  --------
      163.334µs     163.334µs     163.334µs          0        0       0      1  index.html
       23.749µs       5.937µs      19.916µs         25       50       2      4  partials/dynamic1.html
        9.625µs       4.812µs        6.75µs        100       50       1      2  partials/static1.html
        7.625µs       7.625µs       7.625µs        100        0       0      1  partials/static2.html
```

My statement about _A cached template gets executed once before it gets cached, so the "percent cached" will never be 100_ could be questioned ...

An example of that would be `partials/static1.html` above, which is included twice with `partialCached`.  So the cached percentage is 50. In real situations this will probably round up to 100, so I guess this is correct. Just wanted to shout out about it.