@moorereason 2022-02-16
I agree with your comments @bep and the revisions (at least from looking at it on my phone). Would it be clearer if the "cache count" column was called "cache hits"?